User and Restaurant Controller.

+++ src/main/java/com/baduwal/ecommerce/data/entity/Order.java
@@ -10,10 +10,10 @@ public class Order {
 
     private final List<CartItem> cartItems;
     private final int orderId;
-    private final int userId;
+    private final Long userId;
     private final OrderStatus orderStatus;
 
-    public Order(List<CartItem> cartItems, int orderId, int userId, OrderStatus orderStatus) {
+    public Order(List<CartItem> cartItems, int orderId, Long userId, OrderStatus orderStatus) {
         this.cartItems = cartItems;
         this.orderId = orderId;
         this.userId = userId;
